Moved AATAPipeline into extracters.aata_data.py.
